Added alternate (better) version of depthtolaserscan
This adds a second version (depthtolaserscan2.cpp) which handles
contacts that fall above or below vertical bounds more correctly.
That is, they are simply ignored and freespace (no obstacle) is
declared only if there is at least one in-bound sample that reads a
valid distance past the max range.

Previously any samples above these bounds were considered equivalent
to in bound samples that read farther than the maximum.  I.e. either
could trigger a freespace reading.  Unfortunately, when an obstacle
gets too close to register correctly, all the vertical in-band samples
read zero (error).  But there can be some out-of-band samples that
still read an actual distance.  So the previous version would report
these (erroneously) as free space, which made it looks like there was
no object in front of the camera.