Moved the alternate depthtolaserscan to default
This basically copies the new version (depthtolaserscan2.cpp) over the
original (depthtolaserscan.cpp), so that everyone will automatically
get the better version (after an updated).  The new version seems
stable enough to push.

This does leave the v2 version (as not to pull the plug for teams that
updated their launch file).  Exactly the same as the original.  Ugly.
But, oh well.